+++ http/http_handlers/src/test/kotlin/com/hexagonkt/http/handlers/HandlerBuilderTest.kt
@@ -5,7 +5,6 @@ import com.hexagonkt.http.model.NOT_FOUND_404
 import com.hexagonkt.http.model.HttpMethod.*
 import com.hexagonkt.http.model.*
 import com.hexagonkt.http.patterns.LiteralPathPattern
-import com.hexagonkt.http.handlers.examples.send
 import com.hexagonkt.http.model.HttpRequest
 import kotlin.test.Test
 import kotlin.test.assertEquals
