+++ http/http_handlers/src/test/kotlin/com/hexagonkt/http/handlers/RoutingTest.kt
@@ -1,10 +1,8 @@
-package com.hexagonkt.http.handlers.examples
+package com.hexagonkt.http.handlers
 
 import com.hexagonkt.core.require
 import com.hexagonkt.http.model.Header
 import com.hexagonkt.http.model.HttpMethod.GET
-import com.hexagonkt.http.handlers.PathHandler
-import com.hexagonkt.http.handlers.path
 import kotlin.test.Test
 import kotlin.test.assertEquals
