+++ http/http_handlers/src/test/kotlin/com/hexagonkt/http/handlers/ServerSentEventsTest.kt
@@ -1,12 +1,10 @@
-package com.hexagonkt.http.handlers.examples
+package com.hexagonkt.http.handlers
 
 import com.hexagonkt.core.fail
 import com.hexagonkt.core.logging.info
 import com.hexagonkt.core.media.TEXT_EVENT_STREAM
 import com.hexagonkt.http.model.HttpMethod.GET
 import com.hexagonkt.http.model.OK_200
-import com.hexagonkt.http.handlers.PathHandler
-import com.hexagonkt.http.handlers.path
 import com.hexagonkt.http.model.ServerEvent
 import kotlin.test.Test
 import java.util.concurrent.Flow
